Changed int to uint64_t on ffnet agent freq ranges.

+++ libgeopm/src/RegionHintRecommender.cpp
@@ -20,26 +20,26 @@
 namespace geopm
 {
     std::unique_ptr<RegionHintRecommender> RegionHintRecommender::make_unique(const std::string &fmap_path,
-                                                                              int min_freq,
-                                                                              int max_freq)
+                                                                              uint64_t min_freq,
+                                                                              uint64_t max_freq)
     {
         return geopm::make_unique<RegionHintRecommenderImp>(fmap_path, min_freq, max_freq);
     }
 
     std::shared_ptr<RegionHintRecommender> RegionHintRecommender::make_shared(const std::string &fmap_path,
-                                                                              int min_freq,
-                                                                              int max_freq)
+                                                                              uint64_t min_freq,
+                                                                              uint64_t max_freq)
     {
         return std::make_shared<RegionHintRecommenderImp>(fmap_path, min_freq, max_freq);
     }
 
-    RegionHintRecommenderImp::RegionHintRecommenderImp(const std::string &fmap_path, int min_freq,
-                                                       int max_freq)
+    RegionHintRecommenderImp::RegionHintRecommenderImp(const std::string &fmap_path,
+                                                       uint64_t min_freq,
+                                                       uint64_t max_freq)
         : m_min_freq(min_freq)
         , m_max_freq(max_freq)
     {
         std::string buf, err;
-
         std::ifstream ffile(fmap_path);
         if (!ffile.is_open()) {
             throw Exception("RegionHintRecommenderImp::" + std::string(__func__) +
